+++ genium/src/test/resources/smoke/free_types/output/swift/smoke/FreeEnumException.swift
@@ -1,0 +1,6 @@
+//
+//
+// Automatically generated. Do not modify. Your changes will be lost.
+import Foundation
+extension FreeEnum : Error {
+}